+++ UtilityMediaPlayer/src/Player.java
@@ -2,7 +2,7 @@
 public interface Player 
 {
 	public void open(String fileName);
-	public void volumeChange(int newVolume);
+	public void volumeChange(double newVolume);
 	public void alternatePlayback();
 	public void skipPlayback();
 	public void changePosition(long playbackPosition);
